fixed create stack already

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: nkietwee <nkietwee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/03 11:09:29 by nkietwee          #+#    #+#             */
-/*   Updated: 2023/03/13 15:26:05 by nkietwee         ###   ########.fr       */
+/*   Updated: 2023/03/22 15:41:18 by nkietwee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,17 +21,18 @@
 //Mandatory
 int		ft_isdigit(char c);
 size_t	ft_strlen(const char*s);
-long		ft_atoi(const char *str);
+long long	ft_atoi(const char *str);
+// long		ft_atoi(const char *str);
 void	ft_putstr_fd(char *s, int fd);
 char	**ft_split(char const *s, char c);
 int		ft_checknum(char **str, int argc);
 
 //Bonus
-//t_list	*ft_lstnew(void *content);
-//void	ft_lstadd_back(t_list **lst, t_list *new);
+// t_list	*ft_lstnew(void *content);
+// void	ft_lstadd_back(t_list **lst, t_list *new);
 // void	ft_lstadd_front(t_list **lst, t_list *new);
-// int		ft_lstsize(t_list *lst);
 // t_list	*ft_lstlast(t_list *lst);
+// int		ft_lstsize(t_list *lst);
 // void	ft_lstdelone(t_list *lst, void (*del)(void *));
 // void	ft_lstclear(t_list **lst, void (*del)(void *));
 // void	ft_lstiter(t_list *lst, void (*f)(void *));
